Added Main Menu Scene
The very first scene we need to make is the main menu scene as all other scenes will branch out from this. For now, we will simply display this scene with an appropriate title.

An abstract Scene class is created that defines the 4 fundamental functions of each scene. Every scene in the game must extend from this base class and provide their own implementation.