Added Achievement Scene + Scene Switching
To experiment with scene switching, I added another scene that will display the achievements that currently just shows a title text just like the main menu.

To switch between scenes, the user clicks on the mouse and the program switches between the main menu and achievement scenes.

A new variable is introduced called 'nextScene' that will store the scene we are about to change into. When the variable is not null, that means we want to enter a new scene, so we set the current scene to the new scene and reset the value of the variable.