Updated Transition Class + NextScene Logic
Previously, each implementation of the Transition class had to also implement the moment the transition reaches halfway and switches scenes, which is a mandatory thing all transitions should do. As such, that functionality is now built into the Transition class, and all a given implementation has to do is define what counts as "halfway" for that transition animation.