Added Abstract Entity Class + Basic Player Class
I have defined a generic class for entities that will contain similar information about anything that can interact with attacks in the level scene. This will include the player, enemies, obstacles, and possibly more.

As a basic first step, I have created the player class and displayed it in the middle of the screen.