Added Rotation Towards Velocity Direction
To experiment with displaying different sprites based on velocity direction, I created a rotation matrix that stores rotation values (in degrees) for each one of the orthogonal and diagonal directions. During display(), the program gets the direction of the target velocity and figures out what the right angle is to rotate towards.

Also now you can only switch scenes with the right mouse button.

Fixed a bug where the horizontal dashing wasn't working.