Added Key Hold Detection + Charged Bullet/Laser
Now the player can hold down a key and the program will detect if it's held down longer than a specific threshold. This is useful to detect the charged attacks of the player.

To utilize this key holding behavior, I have implemented two new attacks: ChargedBullet and Laser. The charged bullet pass through multiple enemies and damage all of them along the way. The number of enemies it passes through depends on how long it was charged. The laser is activated when the bullet reaches maximum charge, and as such it will be able to pass through an unlimited number of enemies.