Added Multiple Path Types for Troops
Previously, I would create subclasses of the TroopGroup to define the different paths, but this didn't logically make sense as the path and the troop itself weren't tied to each other. To decouple this further, I defined a PathGenerator abstract class that simply returns a list of points that a troop group has to follow, and every implementation of this class will have their own unique path. I have added the following paths:
- XPath: goes to the corners of a rectangle as if it was drawing an X
- HorizontalBarsPath: moves across the screen horizontally, and each time it gets to the edge it moves one row down. Once it gets to the bottom it goes back up in the same way.
- VerticalBarsPath: moves up and down the screen, and each time it gets to the edge it moves across one column. Once it gets to the end it goes back the same way.
- HorizontalLineBath: Just moves across the top of the screen horizontally.
- RandomPath: picks 8 random points in a region