Added Level Complete Panel and Animations
Now, once the player finishes all the waves, the ship flies offscreen and a "level complete" panel shows up which will provide all the stats and buttons to update the scene. As of now, it only has one button that closes the panel.

I started the implementation of the p5.touchgui library. A custom panel class was made that can hold GUI objects, which will become active or inactive depending on whether the panel opens or closes.