Added Map Logic + Preload Structure
Previously, the map scene had no levels to interact with. Now, a levelStructures.json file contains all the information needed for each level. Initially, all levels are incomplete and locked, but the first level is unlocked. As the player completes each level, they set it to complete and unlock the next level. This information will be reflected in the map.

Some helper functions have been created to more easily access the levelStructures data.