Added Basics of Story Scenes
StoryScenes now load the appropriate storyInfo.json file that contains information about how the pages of the story are going to be displayed during the scene. Each page has layers (from background to foreground) and the story scene dynamically moves these layers into position to create a 3D and active-looking page.

The player can also hold the 's' key to skip the scene.