Added Interactable Upgrades to Upgrade Menu
The upgrade menu now loads in an upgradeInfo.json file that contains information about each upgrade, such as the name, price to unlock, prices to improve, and icon. The UpgradeScene then creates the necessary UI for each upgrade, allowing the player to either purchase a new upgrade or improve an existing one only if they have enough currency.