Added Obstacle Class, Asteroids, and Spawners
I've created a generic obstacle class that is an entity which moves down the screen and can bump into the player. The first implementation of the obstacle is the asteroid. Each level specifies which obstacles it will contain, and the appropriate spawner classes will be created that each handle their own way of spawning obstacles throughout the level. I have implemented a basic asteroid spawner and asteroid wall spawner in this commit.