Bugfix: Game Completion always Returned to Map
Previously, when a story scene would finish, it would check if the player has completed all levels to determine whether or not to transition to the level scene or the map scene. However, once the player completed the game, every story scene would transition to the map scene. Now, only the last story scene transitions to the map scene.