Added EnemyRushMode and LevelMode
Previously, the LevelScene would be the only way a player could attack enemies and go through the waves/levels. Since, we have two other modes (Enemy Rush and Survival), it makes sense to create a parent GameScene class instead, which does most of the game logic of LevelScene but the specifics of how waves are created and what happens when the player wins can be decided by each class's implementation. For now, I made two classes implement the GameScene, which are LevelMode and EnemyRushMode.