Fixed Player Movement and Ability Logic
Previously, the special ability would be activated by double clicking the space bar, which since it was also used to shoot bullets, it would be mis-pressed often. Now, the special ability is activated by pressing the 's' key.

Also, player dodging and switching between opposite arrows would stop the player in place even if the user is pressing down on an arrow key, so now the movement is much more fluid and connected.