Panel Buttons now Work Properly in Different Modes
Previously, setting the logic of the retry and map/home buttons of the game panels all required recreating the button objects in each implementation of the GameScene.

Now, the GameScene creates all the buttons as normal, and adds them to the appropriate panels. But instead, each subclass calls implementPanelButtonLogic with their own dictionary that describes the functions that will be called for each button press. If a button label is not present in this dictionary, it will be removed from the panel. This allows for flexibility as the only thing the subclasses need to worry about is the logic of the buttons, not where they are placed.