Added Screen Scaling to Match Canvas Dimensions
Previously, changing the size of the window would adjust the width and height of the sketch, but the positions and sizes of all entities and UI would remain absolute so it would mess up the game if the window resized.

Now, the sketch simply scales itself based on the ratio of old resolution to new resolution. This works just fine for anything natively drawn in p5, but the GUI objects needed some extra code to handle scaling their position, size, and styling properties according to the new resolution. A standard width W and height H has been declared which represents the base size I view on my big monitor, and this is used as a reference to figure out where everything should be drawn on the screen (instead of using p5's built in "width" and "height")